Extracted application root widget to separate file to remove responsibility from main.dart
